+++ class/techatm.class.php
@@ -1,12 +1,27 @@
 <?php
-
-/**
- * Class TechATM
- * Class utilisée pour des mises à jours technique du module
+/* Copyright (C) 2025 ATM Consulting support@atm-consulting.fr
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see https://www.gnu.org/licenses/.
  */
+
 
 namespace userNavHistory;
 
+/**
+ * Class TechATM
+ * Handle some ATM technical features
+ */
 class TechATM
 {
 
@@ -58,17 +73,23 @@ class TechATM
 	/**
 	 *  Constructor
 	 *
-	 *  @param DoliDB $db
+	 * @param DoliDB $db Database handler
+	 * @return void
 	 */
-	function __construct($db)
+	public function __construct($db)
 	{
 		$this->db = $db;
 	}
 
 	/**
-	 * @param DolibarrModules $moduleDescriptor
+	 * Retrieve the About page
+	 *
+	 * @param  DolibarrModules $moduleDescriptor  descriptor module
+	 * @param  bool            $useCache          Use cache
+	 * @return false|string                       display HTML or false
 	 */
-	function getAboutPage($moduleDescriptor, $useCache = true){
+	public function getAboutPage($moduleDescriptor, $useCache = true)
+	{
 		global $langs;
 
 		$url = self::ATM_TECH_URL.'/modules/modules-page-about.php';
@@ -82,11 +103,11 @@ class TechATM
 		$cacheFileName = dol_sanitizeFileName($moduleDescriptor->name.'_'.$langs->defaultlang).'.html';
 		$cacheFilePath = $cachePath.'/'.$cacheFileName;
 
-		if($useCache && is_readable($cacheFilePath)){
+		if ($useCache && is_readable($cacheFilePath)) {
 			$lastChange = filemtime($cacheFilePath);
-			if($lastChange > time() - 86400){
+			if ($lastChange > time() - 86400) {
 				$content = @file_get_contents($cacheFilePath);
-				if($content !== false){
+				if ($content !== false) {
 					return $content;
 				}
 			}
@@ -94,7 +115,7 @@ class TechATM
 
 		$content = $this->getContents($url);
 
-		if(!$content){
+		if (!$content) {
 			$content = '';
 			// About page goes here
 			$content.= '<div style="float: left;"><img src="../img/Dolibarr_Preferred_Partner_logo.png" /></div>';
@@ -104,14 +125,14 @@ class TechATM
 			$content.= '</center>';
 		}
 
-		if($useCache){
-			if(!is_dir($cachePath)){
+		if ($useCache) {
+			if (!is_dir($cachePath)) {
 				$res = dol_mkdir($cachePath, DOL_DATA_ROOT);
-			}else{
+			} else {
 				$res = true;
 			}
 
-			if($res){
+			if ($res) {
 				$comment = '<!-- Generated from '.$url.' -->'."\r\n";
 
 				file_put_contents(
@@ -125,18 +146,11 @@ class TechATM
 	}
 
 	/**
-	 * @param string $moduleTechMane
+	 * @param DolibarrModules $moduleDescriptor Module descriptor instance
+	 * @return string
 	 */
-	public static function getModuleDocUrl($moduleTechMane){
-		$url = self::ATM_TECH_URL.'/modules/doc-redirect.php';
-		$url.= '?module='.$moduleTechMane;
-		return $url;
-	}
-
-	/**
-	 * @param DolibarrModules $moduleDescriptor
-	 */
-	public static function getLastModuleVersionUrl($moduleDescriptor){
+	public static function getLastModuleVersionUrl($moduleDescriptor)
+	{
 		$url = self::ATM_TECH_URL.'/modules/modules-last-version.php';
 		$url.= '?module='.$moduleDescriptor->name;
 		$url.= '&number='.$moduleDescriptor->numero;
@@ -146,192 +160,48 @@ class TechATM
 		return $url;
 	}
 
-
 	/**
-	 * @param $url
-	 * @return false|object
+	 * Fetches URL content using Dolibarr native function.
+	 *
+	 * @param string $url The URL to fetch.
+	 * @return array|false The fetched content, or false on failure.
 	 */
-	public function getJsonData($url){
-		$this->data = false;
-		$res = @file_get_contents($url);
-		$this->http_response_header = $http_response_header;
-		$this->TResponseHeader = self::parseHeaders($http_response_header);
-		if($res !== false){
-			$pos = strpos($res, '{');
-			if($pos > 0){
-				// cela signifie qu'il y a une erreur ou que la sortie n'est pas propre
-				$res = substr($res, $pos);
-			}
-
-			$this->data = json_decode($res);
-		}
-
-		return $this->data;
-	}
-
-	/**
-	 * @param $url
-	 * @return false|string
-	 */
-	public function getContents($url){
-		$this->data = false;
-		$res = @file_get_contents($url);
-		$this->http_response_header = $http_response_header;
-		$this->TResponseHeader = self::parseHeaders($http_response_header);
-		if($res !== false){
-			$this->data = $res;
-		}
-		return $this->data;
-	}
-
-	public static function http_response_code_msg($code = NULL)
+	public function getContents($url)
 	{
-		if ($code !== NULL) {
+		global $conf;
+		require_once DOL_DOCUMENT_ROOT . '/core/lib/files.lib.php';
 
-			switch ($code) {
-				case 100:
-					$text = 'Continue';
-					break;
-				case 101:
-					$text = 'Switching Protocols';
-					break;
-				case 200:
-					$text = 'OK';
-					break;
-				case 201:
-					$text = 'Created';
-					break;
-				case 202:
-					$text = 'Accepted';
-					break;
-				case 203:
-					$text = 'Non-Authoritative Information';
-					break;
-				case 204:
-					$text = 'No Content';
-					break;
-				case 205:
-					$text = 'Reset Content';
-					break;
-				case 206:
-					$text = 'Partial Content';
-					break;
-				case 300:
-					$text = 'Multiple Choices';
-					break;
-				case 301:
-					$text = 'Moved Permanently';
-					break;
-				case 302:
-					$text = 'Moved Temporarily';
-					break;
-				case 303:
-					$text = 'See Other';
-					break;
-				case 304:
-					$text = 'Not Modified';
-					break;
-				case 305:
-					$text = 'Use Proxy';
-					break;
-				case 400:
-					$text = 'Bad Request';
-					break;
-				case 401:
-					$text = 'Unauthorized';
-					break;
-				case 402:
-					$text = 'Payment Required';
-					break;
-				case 403:
-					$text = 'Forbidden';
-					break;
-				case 404:
-					$text = 'Not Found';
-					break;
-				case 405:
-					$text = 'Method Not Allowed';
-					break;
-				case 406:
-					$text = 'Not Acceptable';
-					break;
-				case 407:
-					$text = 'Proxy Authentication Required';
-					break;
-				case 408:
-					$text = 'Request Time-out';
-					break;
-				case 409:
-					$text = 'Conflict';
-					break;
-				case 410:
-					$text = 'Gone';
-					break;
-				case 411:
-					$text = 'Length Required';
-					break;
-				case 412:
-					$text = 'Precondition Failed';
-					break;
-				case 413:
-					$text = 'Request Entity Too Large';
-					break;
-				case 414:
-					$text = 'Request-URI Too Large';
-					break;
-				case 415:
-					$text = 'Unsupported Media Type';
-					break;
-				case 500:
-					$text = 'Internal Server Error';
-					break;
-				case 501:
-					$text = 'Not Implemented';
-					break;
-				case 502:
-					$text = 'Bad Gateway';
-					break;
-				case 503:
-					$text = 'Service Unavailable';
-					break;
-				case 504:
-					$text = 'Gateway Time-out';
-					break;
-				case 505:
-					$text = 'HTTP Version not supported';
-					break;
-				default:
-					$text = 'Unknown http status code "' . htmlentities($code) . '"';
-					break;
-			}
-
-			return $text;
-
-		} else {
-			return $text = 'Unknown http status code NULL';
+		$content = getURLContent($url, 'GET', '', 1, 5);
+		if ($content !== false) {
+			$this->data = $content;
+			return $this->data;
 		}
-	}
 
-	public static function parseHeaders( $headers )
+		return false;
+	}
+	/**
+	 * Parses raw HTTP header lines into an associative array.
+	 *
+	 * @param array $headers List of raw header strings.
+	 * @return array Key-value headers, including the 'reponse_code'.
+	 */
+	public static function parseHeaders($headers)
 	{
 		$head = array();
-		if(!is_array($headers)){
+		if (!is_array($headers)) {
 			return $head;
 		}
 
-		foreach( $headers as $k=>$v )
-		{
-			$t = explode( ':', $v, 2 );
-			if( isset( $t[1] ) )
-				$head[ trim($t[0]) ] = trim( $t[1] );
-			else
-			{
+		foreach ($headers as $k=>$v) {
+			$t = explode(':', $v, 2);
+			if ( isset($t[1]) )
+				$head[ trim($t[0]) ] = trim($t[1]);
+			else {
 				$head[] = $v;
-				if( preg_match( "#HTTP/[0-9\.]+\s+([0-9]+)#",$v, $out ) )
+				if ( preg_match("#HTTP/[0-9\.]+\s+([0-9]+)#", $v, $out) )
 					$head['reponse_code'] = intval($out[1]);
 			}
 		}
 		return $head;
 	}
-
 }